requirements.txt
argparse~=1.4.0
python-dotenv~=0.21.0
requests~=2.28.1
pandas~=1.5.2